+++ leda/components/DropZone/theme.ts
@@ -9,9 +9,9 @@ export const defaultDropZoneTheme = {
   fileDeleteLink:       'dropzone-file-delete internal',
   fileDownloadLink:     'dropzone-file-download internal',
   invalid:              'danger',
+
   rejectedFilesWrapper: 'rejected-list margin-none',
   rejectedMessage:      'dropzone-rejected-message txt-danger',
-
   required:             'required',
   status:               'dropzone-status',
   wrapper:              'dropzone-wrapper',