+++ CHANGELOG.md
@@ -8,11 +8,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Tour: Added stepDelay prop, added positions: top-left, bottom-left, top-center, bottom-center
+- DropZone: Added isLoading, loadingProgress and loadingViewRender props
 
-### Fixed 
+### Fixed
 - Validation: Fixed displaying of invalidMessage with isValid prop set to false
 - NumericRange: Fixed null value display bug
 - Tour: Added class 'tour-active-element' for active element to customize styles
+- FileDrop, DropZone: Added loadingText prop
 
 
 ## [0.34.0] - 2020-11-09